api/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
SQLAlchemy==2.0.35
pydantic==2.9.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
